latest

+++ html/Uri.cc
@@ -328,7 +328,7 @@ static unsigned short default_port_for_scheme(const char *scheme_str)
 
 	for (scheme = schemes; scheme->name != NULL; ++scheme)
 		if (strcasecmp(scheme_str, scheme->name) == 0)
-			return scheme->default_port;
+			return (unsigned short)scheme->default_port;
 
 	return 0;
 }
